+++ src/pytgpt/__init__.py
@@ -1,53 +1,24 @@
 from .utils import appdir
+import g4f
 
-__version__ = "0.3.1"
+__version__ = "0.3.2"
 __author__ = "Smartwa"
 __repo__ = "https://github.com/Simatwa/python-tgpt"
 
-gpt4free_providers = [
-    "AItianhuSpace",
-    "AiChatOnline",
-    "Aura",
-    "Bard",
-    "Bing",
-    "ChatBase",
-    "ChatForAi",
-    "Chatgpt4Online",
-    "ChatgptAi",
-    "ChatgptDemo",
-    "ChatgptNext",
-    "Chatxyz",
-    "DeepInfra",
-    "FakeGpt",
-    "FreeChatgpt",
-    "GPTalk",
-    "GeekGpt",
-    "GeminiProChat",
-    "Gpt6",
-    "GptChatly",
-    "GptForLove",
-    "GptGo",
-    "GptTalkRu",
-    "Hashnode",
-    "HuggingChat",
-    "Koala",
-    "Liaobots",
-    "Llama2",
-    "MyShell",
-    "OnlineGpt",
-    "OpenaiChat",
-    "PerplexityAi",
-    "Phind",
-    "Pi",
-    "Poe",
-    "Raycast",
-    "TalkAi",
-    "Theb",
-    "ThebApi",
-    "You",
-    "Yqcloud",
+tgpt_providers = [
+    "leo",
+    "openai",
+    "fakeopen",
+    "opengpt",
+    "koboldai",
+    "bard",
 ]
-available_providers = ["leo", "openai", "fakeopen", "opengpt", "koboldai", "bard"] + gpt4free_providers
+
+gpt4free_providers = [
+    provider.__name__ for provider in g4f.Provider.__providers__  # if provider.working
+]
+
+available_providers = tgpt_providers + gpt4free_providers
 
 __all__ = [
     "appdir",