+++ docs/CHANGELOG.md
@@ -245,3 +245,13 @@ Assume role of the LLM and give your response.*
 
 - **Aura** is the default provider
 - Other minor fixes.
+
+## v0.3.2
+
+**What's new?**
+
+- Add utility commands `utils` `gpt4free`  `update` etc
+- Support g4f-based non-working providers. 
+- Drop restriction to specific `g4f` dependency version (https://github.com/Simatwa/python-tgpt/issues/14#issuecomment-1899468911)
+- Pass auth value to g4f-based providers. `-k`
+- Support proxying in g4f providers